js: load only URL settings on URL change
When a user clicks Back or Forward button in the browser, handler
`window.onpopstate` is called.  Handler set up in bingo.js loads all
settings.  For URL settings, this means new sheet generation, which
possibly includes highlighting of the Never goals.  For localStorage
settings this means colour theme loading, which includes a smooth
transition.  This means that highlight of Never goals is also smooth,
which is undesirable.

Settings stored in localStorage don't need to be loaded when URL
changes.  So just load only the URL settings to 1) avoid the undesirable
animation for highlight of Never goals and 2) avoid the unnecessary
re-loading of local settings.